requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.25
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.26.0

